feat: integrate backend login form into layout
- Move login form from separate page to backend layout component
- Update middleware to allow unauthenticated access only to /backend root
- Protect all /backend/* sub-routes requiring authentication
- Add error messaging for buyers attempting backend access
- Remove obsolete /backend/login page
- Maintain logout functionality in navigation bar

Security: Buyers can see login form but cannot access backend features
