+++ middleware.ts
@@ -7,7 +7,9 @@ export async function middleware(req: NextRequest) {
   const isBackend = req.nextUrl.pathname.startsWith("/backend");
   
   // Basic auth check in middleware
-  if (isBackend && !token) {
+  // Allow /backend root without token (for login form)
+  // But protect all sub-routes like /backend/tickets, /backend/buyers, etc.
+  if (isBackend && !token && req.nextUrl.pathname !== "/backend") {
     const url = new URL("/auth/signin", req.url);
     url.searchParams.set("callbackUrl", req.nextUrl.pathname + req.nextUrl.search);
     return NextResponse.redirect(url);
