--- src/app/backend/login/page.tsx
@@ -1,72 +1,0 @@
-"use client";
-
-import { useState } from "react";
-import { signIn, useSession } from "next-auth/react";
-
-export default function TicketReservesPage() {
-  const { status } = useSession();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleSignIn = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    try {
-      await signIn("credentials", { email, password, callbackUrl: "/backend" });
-    } catch (error) {
-      console.error("Sign in error:", error);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  if (status === "loading") {
-    return <div className="p-8">Loading...</div>;
-  }
-
-
-  return (
-    <div className="p-8 max-w-md mx-auto">
-      <h1 className="text-2xl font-bold mb-6">Einloggen</h1>
-      <form onSubmit={handleSignIn} className="space-y-4">
-        <div>
-          <label htmlFor="email" className="block text-sm font-medium mb-2">
-            Email Adresse
-          </label>
-          <input
-            id="email"
-            type="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-            className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-            placeholder="Email Adresse"
-          />
-        </div>
-        <div>
-          <label htmlFor="password" className="block text-sm font-medium mb-2">
-            Passwort
-          </label>
-          <input
-            id="password"
-            type="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-            className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-            placeholder="Passwort"
-          />
-        </div>
-        <button
-          type="submit"
-          disabled={isLoading}
-          className="w-full bg-blue-500 text-white py-2 px-4 rounded-md hover:bg-blue-600 disabled:opacity-50"
-        >
-          {isLoading ? "Übertrage..." : "Einloggen"}
-        </button>
-      </form>
-        {/* <p className="mt-4 text-sm text-gray-600">Bitte mit Email und Passwort anmelden.</p> */}
-    </div>
-  );
-}
